requirements.txt
PyMySQL==1.1.0
python-dotenv==1.0.0
types-PyMySQL==1.1.0.1
